generate-image/requirements.txt
asn1crypto==1.5.1
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
install==1.3.5
minio==7.1.12
pg8000==1.29.3
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
scramp==1.4.4
six==1.16.0
testing.common.database==2.0.3
testing.postgresql==1.3.0
urllib3==1.26.12
